Add trx pointer to struct mtr_t
Summary:
Feature: Basic Code Changes

This adds a pointer to the trx to each mtr.
This allows the trx to be accessed in parts of the code
where it was otherwise not available.

Other than adding the extra field in the structure,
and filling it properly, this change, on its own, should
have no effect.  It is basically a non-change.

This is a base change required for lots of stuff,
including innodb_fake_changes, table stats, etc....

I am factoring this out to allow diff re-ordering.

Test Plan:
Build is clean, and so are mtr tests.

This code was factored out of the existing table stats diff.
There are no overall code changes once table stats diff is applied.

This change has	been running in	production for about a year now.
All is still well, so this is pretty safe.

Reviewers: pivanof, inaam-rana, liang.guo.752

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan, flamingcow, jeremycole, andrew-ford

Differential Revision: https://reviews.facebook.net/D16209